+++ src/main/java/com/telepathicgrunt/the_bumblezone/tags/BZBlockTags.java
@@ -2,8 +2,11 @@ package com.telepathicgrunt.the_bumblezone.tags;
 
 import com.telepathicgrunt.the_bumblezone.Bumblezone;
 import net.minecraft.block.Block;
+import net.minecraft.entity.EntityType;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.EntityTypeTags;
 import net.minecraft.tags.ITag;
+import net.minecraft.util.ResourceLocation;
 
 public class BZBlockTags {
     // All tag wrappers need to be made at mod init.
@@ -14,4 +17,6 @@ public class BZBlockTags {
     public static final ITag.INamedTag<Block> BLACKLISTED_RESOURCEFUL_COMBS = BlockTags.makeWrapperTag(Bumblezone.MODID+":blacklisted_resourceful_bees_combs");
     public static final ITag.INamedTag<Block> BLACKLISTED_PRODUCTIVEBEES_COMBS = BlockTags.makeWrapperTag(Bumblezone.MODID+":blacklisted_productive_bees_combs");
     public static final ITag.INamedTag<Block> HONEYCOMBS_THAT_FEATURES_CAN_CARVE = BlockTags.makeWrapperTag(Bumblezone.MODID+":honeycombs_that_features_can_carve");
+
+    public static final ITag.INamedTag<EntityType<?>> BLACKLISTED_RESOURCEFUL_BEES_ENTITIES = EntityTypeTags.createOptional(new ResourceLocation(Bumblezone.MODID, "blacklisted_resourceful_bees_entities"));
 }